update metadata variables

+++ src/main/java/dk/kb/labsapi/metadataFormats/FullPageMetadata.java
@@ -4,20 +4,20 @@ import org.apache.solr.client.solrj.response.QueryResponse;
 
 public class FullPageMetadata extends BasicMetadata {
     private String pageUUID;
-    private double pageWidth;
-    private double pageHeight;
+    private Long pageWidth;
+    private Long pageHeight;
 
-    public FullPageMetadata(String pageUUID, double pageWidth, double pageHeight){
+    public FullPageMetadata(String pageUUID, Long pageWidth, Long pageHeight){
         this.pageUUID = pageUUID;
         this.pageHeight = pageHeight;
         this.pageWidth = pageWidth;
     }
 
-    public void setPageHeight(double pageHeight) {
+    public void setPageHeight(Long pageHeight) {
         this.pageHeight = pageHeight;
     }
 
-    public void setPageWidth(double pageWidth) {
+    public void setPageWidth(Long pageWidth) {
         this.pageWidth = pageWidth;
     }
